add apollo call to Repository add MainActivityViewModel remove notification and saved bottom nav tabs add MainActivity and fragments to view package folder
