+++ app/src/main/java/com/lambda_labs/community_calendar/App.kt
@@ -20,5 +20,6 @@ class App: Application() {
         token = sharedPrefs.getString(TOKEN_KEY, "")
 
         repository = Repository(this)
+
     }
 }
