+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeFragment">
+    tools:context=".view.HomeFragment">
 
     <com.google.android.material.textview.MaterialTextView
         android:id="@+id/txt_title_featured"
@@ -28,8 +28,7 @@
         android:text="@string/see_all"
         android:textSize="14sp"
         app:layout_constraintBottom_toTopOf="@id/featured_event_recycler"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/search_bar" />
+        app:layout_constraintEnd_toEndOf="parent" />
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/featured_event_recycler"
