+++ app/src/main/res/menu/bottom_nav.xml
@@ -7,19 +7,9 @@
         android:title="@string/home" />
 
     <item
-        android:id="@+id/notification_icon"
-        android:icon="@drawable/notification_icon"
-        android:title="@string/notifications" />
-
-    <item
         android:id="@+id/event_icon"
         android:icon="@drawable/events_icon"
         android:title="@string/events" />
-
-    <item
-        android:id="@+id/saved_icon"
-        android:icon="@drawable/saved_icon"
-        android:title="@string/saved" />
 
     <item
         android:id="@+id/loginFragment"
